when dealing with integers that can go higher than 65535, then a long must be used instead of int
git-svn-id: https://svn.code.sf.net/p/gopherus/code/trunk@18 767d67df-c2ba-418b-950d-eab06a56d38d

+++ int2str.h
@@ -6,5 +6,5 @@
 #ifndef int2str_h_sentinel
   #define int2str_h_sentinel
   /* converts an integer from the range 0..9999999 into a string. returns the length of the computed string, or -1 on error. */
-  int int2str(char *res, int x);
+  int int2str(char *res, long x);
 #endif
